# backend/src/main/java/com/nexus/agent/api/dto/GetRunResponse.java
package com.nexus.agent.api.dto;

import java.time.Instant;

public record GetRunResponse(
        String runId,
        String sessionId,
        String status,
        Instant createdAt,
        Instant updatedAt
) {}