# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus Agent API
  version: 0.2.0
  description: M2 最小可用契约（Session / Run / Event / Approval）。

servers:
  - url: http://localhost:8080
    description: 本地后端

paths:
  /api/v1/sessions:
    post:
      summary: 创建会话
      operationId: createSession
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"

  /api/v1/sessions/{sessionId}/runs:
    post:
      summary: 提交一次运行（M2 仍可为空任务）
      operationId: createRun
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: 已创建运行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRunResponse"
        "404":
          description: 会话不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

  /api/v1/runs/{runId}/events:
    get:
      summary: 订阅运行事件（SSE，text/event-stream）
      operationId: subscribeRunEvents
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE 流
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: 运行不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

  /api/v1/runs/{runId}:
    get:
      summary: 查询运行状态
      operationId: getRun
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRunResponse"
        "404":
          description: 运行不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

  /api/v1/runs/{runId}/approvals:
    post:
      summary: 创建审批请求
      operationId: createApproval
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRequest"
      responses:
        "201":
          description: 已创建审批
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
        "404":
          description: 运行不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

  /api/v1/approvals/{approvalId}:
    get:
      summary: 查询审批状态
      operationId: getApproval
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
        "404":
          description: 审批不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

  /api/v1/approvals/{approvalId}/resolve:
    post:
      summary: 确认审批（同意/拒绝）
      operationId: resolveApproval
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveApprovalRequest"
      responses:
        "200":
          description: 确认成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
        "400":
          description: 入参非法或重复确认
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"
        "404":
          description: 审批不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

components:
  schemas:
    CreateSessionResponse:
      type: object
      required: [sessionId, createdAt]
      properties:
        sessionId:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateRunResponse:
      type: object
      required: [runId, sessionId, status, createdAt, updatedAt]
      properties:
        runId:
          type: string
        sessionId:
          type: string
        status:
          type: string
          description: 运行状态（PENDING/STREAMING/COMPLETED）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    GetRunResponse:
      allOf:
        - $ref: "#/components/schemas/CreateRunResponse"

    CreateApprovalRequest:
      type: object
      properties:
        action:
          type: string
          description: 审批动作名称，默认 WRITE_OPERATION

    ResolveApprovalRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        resolvedBy:
          type: string
          description: 确认人，默认 system

    ApprovalResponse:
      type: object
      required: [approvalId, runId, sessionId, action, status, createdAt]
      properties:
        approvalId:
          type: string
        runId:
          type: string
        sessionId:
          type: string
        action:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        resolvedBy:
          type: string
          nullable: true

    ApiErrorBody:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        runId:
          type: string
          nullable: true
        stepId:
          type: string
          nullable: true